Add support for setting LIBDIR via env
Hardcoding the libdir as `${prefix}/lib` isn't correct for systems
that keep 64-bit libraries in `${prefix}/lib64`.

Allow setting a variable `LIBDIR` in the environment, to define both
the pkg-config libdir and the path passed to cdylib_link_lines().

`LIBDIR` may be an absolute path or one based on `${prefix}`, which
will be substituted before creating the PathBuf for cdylib_link_lines.