tests: build lib for symbols tests conditionally
When the tests are executed, cargo might not be available,
however the library could be built already.

An example of when this could happen is cross-compiling for
embedded systems: when the library is compiled for a different
(potentially low powered) system, then at runtime frequently
there is no rust toolchain available to build the library;
however the library is usually already built and installed.

To be able to run the test on pre-built artifacts, introduce
an environment variable, "TEST_DONT_BUILD_LIB", which, if
present with any value, will skip building the library, thus
allow running the tests without having a Rust toolchain installed.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>